extract tracker blocking script
