+++ Core/WKWebViewConfigurationExtension.swift
@@ -78,104 +78,4 @@ extension WKWebViewConfiguration {
             self.userContentController.add(rulesList)
         }
     }
-
-    /*
-    public func loadScripts(storageCache: StorageCache, contentBlockingEnabled: Bool) {
-        Loader(contentController: userContentController,
-               storageCache: storageCache,
-               injectContentBlockingScripts: contentBlockingEnabled).load()
-    }
-    */
 }
-
-/*
-private struct Loader {
-
-    struct CacheNames {
-
-        static let surrogateJson = "surrogateJson"
-
-    }
-    
-    let cache = ContentBlockerStringCache()
-    let javascriptLoader = JavascriptLoader()
-
-    let userContentController: WKUserContentController
-    let injectContentBlockingScripts: Bool
-    
-    let whitelist: String
-    let surrogates: String
-    let trackerData: String
-
-    init(contentController: WKUserContentController, storageCache: StorageCache, injectContentBlockingScripts: Bool) {
-        self.userContentController = contentController
-        self.injectContentBlockingScripts = injectContentBlockingScripts
-        
-        self.whitelist = (WhitelistManager().domains?.joined(separator: "\n") ?? "")
-            + "\n"
-            + (storageCache.fileStore.loadAsString(forConfiguration: .temporaryWhitelist) ?? "")
-        self.surrogates = storageCache.fileStore.loadAsString(forConfiguration: .surrogates) ?? ""
-
-        // Encode whatever the tracker data manager is using to ensure it's in sync and because we know it will work
-        let encodedTrackerData = try? JSONEncoder().encode(TrackerDataManager.shared.trackerData)
-        self.trackerData = String(data: encodedTrackerData!, encoding: .utf8)!
-    }
-
-    func load() {
-        let spid = Instruments.shared.startTimedEvent(.injectScripts)
-        loadMainFrameOnlyScripts()
-
-        loadMessagingScripts()
-        
-        if injectContentBlockingScripts {
-            loadContentBlockingScripts()
-        }
-        
-        loadLoginDetection()
-
-        Instruments.shared.endTimedEvent(for: spid)
-    }
-
-    private func loadLoginDetection() {
-        load(scripts: [ .loginDetection ], forMainFrameOnly: false)
-    }
-
-    private func loadMainFrameOnlyScripts() {
-        if #available(iOS 13, *) {
-            load(scripts: [ .findinpage  ] )
-        } else {
-            load(scripts: [ .document, .findinpage ] )
-        }
-    }
-
-    private func loadContentBlockingScripts() {
-        javascriptLoader.load(script: .contentblocker, withReplacements: [
-            "${whitelist}": whitelist,
-            "${trackerData}": trackerData,
-            "${surrogates}": surrogates
-        ], into: userContentController, forMainFrameOnly: false)
-        load(scripts: [ .detection ], forMainFrameOnly: false)
-    }
-
-    private func loadMessagingScripts() {
-        load(scripts: [ .messaging ], forMainFrameOnly: false)
-
-        if isDebugBuild {
-            javascriptLoader.load(script: .debugMessagingEnabled,
-                                  into: userContentController,
-                                  forMainFrameOnly: false)
-        } else {
-            javascriptLoader.load(script: .debugMessagingDisabled,
-                                  into: userContentController,
-                                  forMainFrameOnly: false)
-        }
-    }
-
-    private func load(scripts: [JavascriptLoader.Script], forMainFrameOnly: Bool = true) {
-        for script in scripts {
-            javascriptLoader.load(script, into: userContentController, forMainFrameOnly: forMainFrameOnly)
-        }
-    }
-
-}
-*/
