--- Core/messaging.js
@@ -1,44 +1,0 @@
-//
-//  messaging.js
-//  DuckDuckGo
-//
-//  Copyright © 2017 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-var duckduckgoMessaging = function() {
-
-	function trackerDetected(data) {
-		try {
-			webkit.messageHandlers.trackerDetectedMessage.postMessage(data);
-		} catch(error) {
-			// webkit might not be defined
-		}
-	}
- 
-    function loginFormDetected() {
-        try {
-            webkit.messageHandlers.loginFormDetected.postMessage({});
-        } catch(error) {
-            // webkit might not be defined
-        }
-    }
-
-	return {
-
-		trackerDetected: trackerDetected,
-        loginFormDetected: loginFormDetected
-
-	}
-}()
